Add datagrid empty message tests
